+++ validationframework-core/src/main/java/com/google/code/validationframework/base/property/AbstractReadableWritableProperty.java
@@ -30,8 +30,7 @@ import com.google.code.validationframework.api.property.ReadableWritableProperty
 /**
  * Abstract implementation of a {@link ReadableWritableProperty}.
  * <p/>
- * Sub-classes should call the {@link #notifyListeners(Object, Object)} or {@link #notifyListeners(Object, Object,
- * boolean)} methods whenever the property value changes.
+ * Sub-classes should call the {@link #maybeNotifyListeners(Object, Object)} method whenever the property value changes.
  * <p/>
  * Note that this class is not thread-safe.
  *