+++ validationframework-core/src/main/java/com/google/code/validationframework/base/property/wrap/NegateBooleanPropertyWrapper.java
@@ -50,7 +50,7 @@ public class NegateBooleanPropertyWrapper extends AbstractReadableProperty<Boole
         public void valueChanged(ReadableProperty<Boolean> property, Boolean oldValue, Boolean newValue) {
             Boolean transformedOldValue = transformer.transform(oldValue);
             Boolean transformedNewValue = transformer.transform(newValue);
-            notifyListeners(transformedOldValue, transformedNewValue);
+            maybeNotifyListeners(transformedOldValue, transformedNewValue);
         }
     }