+++ validationframework-experimental/src/main/java/com/google/code/validationframework/swing/property/wrap/InvokeLaterWrapper.java
@@ -61,7 +61,7 @@ public class InvokeLaterWrapper<T> extends AbstractReadableWritableProperty<T, T
          */
         @Override
         public void run() {
-            notifyListeners(oldValue, getValue());
+            maybeNotifyListeners(oldValue, getValue());
         }
     }