+++ validationframework-swing/src/main/java/com/google/code/validationframework/swing/property/JToggleButtonSelectedProperty.java
@@ -124,7 +124,7 @@ public class JToggleButtonSelectedProperty extends AbstractReadableWritablePrope
         if (updatingFromComponent) {
             Boolean oldValue = this.value;
             this.value = value;
-            notifyListeners(oldValue, this.value);
+            maybeNotifyListeners(oldValue, this.value);
         } else if (value != null) {
             component.setSelected(value);
         }